poc/requirements.txt
openai==1.57.0
streamlit==1.41.0
python-dotenv==1.0.1
requests==2.32.3
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
